requirements.txt
mssql-django
Django
python-dotenv
djangorestframework
pillow
django-filter
djangorestframework-simplejwt
requests
django-cors-headers
drf-yasg
redis
flake8
